Fix basket save operation for multiple baskets
When a basket has a predefined ID (for multiple baskets), a `Basket Empty` error is thrown during basket save operation. This occurs because the basket ID is not set in the controller before check operation is done.

This PR fixes the issue by attaching the cart ID to the controller earlier, thereby making sure that subsequent calls to the controller will return the correct basket.

This issue was first posted on https://aimeos.org/help/post13915.html